+++ src/GraphicsControls.Sample/Controls/ColorPicker/ColorPicker.cs
