+++ src/GraphicsControls.Sample/Controls/CustomSlider/CustomSliderDrawable.cs
