+++ src/GraphicsControls.Sample/Controls/CustomSlider/CustomSliderMapper.cs
