+++ src/GraphicsControls.Sample/Controls/CustomSlider/DrawCustomSlider.cs
