+++ src/GraphicsControls.Sample/Controls/Persona/AvatarSize.cs
