+++ src/GraphicsControls.Sample/Controls/Persona/AvatarSizeExtensions.cs
