+++ src/GraphicsControls.Sample/Controls/Persona/Persona.cs
