+++ src/GraphicsControls.Sample/CustomControlsPage.cs
