+++ src/GraphicsControls.Sample/CustomNavigationPage.cs
