+++ src/GraphicsControls.Sample/GalleryPageCSharp.cs
