+++ src/GraphicsControls.Sample/GalleryPageXAML.xaml
