+++ src/GraphicsControls.Sample/GalleryPageXAML.xaml.cs
