--- src/GraphicsControls.Sample/GraphicsControls.Sample.WinUI/GraphicsControls.Sample.WinUI.csproj
@@ -1,47 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-	<PropertyGroup>
-		<TargetFrameworks>net6.0-windows10.0.19041</TargetFrameworks>
-		<RuntimeIdentifier>win-x64</RuntimeIdentifier>
-		<OutputType>WinExe</OutputType>
-		<TargetPlatformMinVersion>10.0.17763.0</TargetPlatformMinVersion>
-		<UseMaui>true</UseMaui>
-		<SingleProject>true</SingleProject>
-		<RootNamespace>GraphicsControls.Sample</RootNamespace>
-
-		<!-- Required - Enable Single Project for WinUI -->
-		<EnablePreviewMsixTooling>true</EnablePreviewMsixTooling>
-		<WindowsPackageType>MSIX</WindowsPackageType>
-	</PropertyGroup>
-
-	<ItemGroup>
-		<!-- App Icon -->
-		<MauiImage Include="..\GraphicsControls.Sample\Resources\appicon.svg" Link="Resources\appicon.svg" ForegroundFile="..\GraphicsControls.Sample\Resources\appiconfg.svg" IsAppIcon="true" Color="#512BD4" />
-
-		<!-- Splash Screen -->
-		<MauiSplashScreen Include="..\GraphicsControls.Sample\Resources\appiconfg.svg" Link="Resources\appiconfg.svg" Color="#512BD4" />
-
-		<!-- Images -->
-		<MauiImage Include="..\GraphicsControls.Sample\Resources\Images\*" Link="Resources\Images\%(Filename)%(Extension)" />
-
-		<!-- Custom Fonts -->
-		<MauiFont Include="..\GraphicsControls.Sample\Resources\Fonts\*" Link="Resources\Fonts\%(Filename)%(Extension)" />
-	</ItemGroup>
-
-	<ItemGroup>
-		<Compile Include="..\GraphicsControls.Sample\**\*.cs" Exclude="..\GraphicsControls.Sample\bin\**;..\GraphicsControls.Sample\obj\**;..\GraphicsControls.Sample\Platforms\**" Link="%(RecursiveDir)%(Filename)%(Extension)" />
-		<MauiXaml Include="..\GraphicsControls.Sample\**\*.xaml" Exclude="..\GraphicsControls.Sample\bin\**;..\GraphicsControls.Sample\obj\**" Link="%(RecursiveDir)%(FileName)%(Extension)" />
-	</ItemGroup>
-
-	<ItemGroup>
-		<PackageReference Include="Microsoft.ProjectReunion" Version="0.8.5" />
-		<PackageReference Include="Microsoft.ProjectReunion.Foundation" Version="0.8.5" />
-		<PackageReference Include="Microsoft.ProjectReunion.WinUI" Version="0.8.5" />
-		<PackageReference Include="Microsoft.ProjectReunion.InteractiveExperiences" Version="0.8.5" NoWarn="NU1701" />
-	</ItemGroup>
-
-	<ItemGroup>
-	  <ProjectReference Include="..\..\GraphicsControls\GraphicsControls.csproj" />
-	</ItemGroup>
-
-</Project>
