--- src/GraphicsControls.Sample/GraphicsControls.Sample.WinUI/Properties/launchSettings.json
@@ -1,7 +1,0 @@
-{
-  "profiles": {
-    "GraphicsControls.Sample.WinUI": {
-      "commandName": "MsixPackage"
-    }
-  }
-}
