--- src/GraphicsControls.Sample/GraphicsControls.Sample/GraphicsControls.Sample.csproj
@@ -1,49 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-	<PropertyGroup>
-		<!-- iOS, Android, MacCatalyst -->
-		<TargetFrameworks>net6.0-ios;net6.0-android;net6.0-maccatalyst</TargetFrameworks>
-		<RuntimeIdentifiers Condition="$(TargetFramework.Contains('-android'))">android-arm;android-arm64;android-x86;android-x64</RuntimeIdentifiers>
-		<OutputType>Exe</OutputType>
-		<UseMaui>true</UseMaui>
-		<SingleProject>true</SingleProject>
-		<RootNamespace>GraphicsControls.Sample</RootNamespace>
-		<EnablePreviewMsixTooling>true</EnablePreviewMsixTooling>
-
-		<!-- Display name -->
-		<ApplicationTitle>GraphicsControls.Sample</ApplicationTitle>
-
-		<!-- App Identifier -->
-		<ApplicationId>com.companyname.GraphicsControls.Sample</ApplicationId>
-
-		<!-- Versions -->
-		<ApplicationVersion>1</ApplicationVersion>
-		<AndroidVersionCode>1</AndroidVersionCode>
-	</PropertyGroup>
-
-	<ItemGroup>
-		<!-- App Icon -->
-		<MauiImage Include="Resources\appicon.svg" ForegroundFile="Resources\appiconfg.svg" IsAppIcon="true" Color="#512BD4" />
-
-		<!-- Splash Screen -->
-		<MauiSplashScreen Include="Resources\appiconfg.svg" Color="#512BD4" />
-
-		<!-- Images -->
-		<MauiImage Include="Resources\Images\*" />
-
-		<!-- Custom Fonts -->
-		<MauiFont Include="Resources\Fonts\*" />
-	</ItemGroup>
-
-	<ItemGroup>
-	  <ProjectReference Include="..\..\GraphicsControls\GraphicsControls.csproj" />
-	</ItemGroup>
-
-	<PropertyGroup>
-		<InvariantGlobalization Condition="$(TargetFramework.Contains('-maccatalyst'))">true</InvariantGlobalization>
-		<RuntimeIdentifier Condition="$(TargetFramework.Contains('-ios'))">iossimulator-x64</RuntimeIdentifier>
-		<RuntimeIdentifier Condition="$(TargetFramework.Contains('-maccatalyst'))">maccatalyst-x64</RuntimeIdentifier>
-		<UseInterpreter Condition="$(TargetFramework.Contains('-android'))">False</UseInterpreter>
-	</PropertyGroup>
-
-</Project>
