+++ src/GraphicsControls.Sample/MainPage.cs
