+++ src/GraphicsControls.Sample/Platforms/Android/MainActivity.cs
@@ -1,0 +1,11 @@
+﻿using Android.App;
+using Android.Content.PM;
+using Microsoft.Maui;
+
+namespace GraphicsControls.Sample
+{
+    [Activity(Theme = "@style/Maui.SplashTheme", MainLauncher = true, ConfigurationChanges = ConfigChanges.ScreenSize | ConfigChanges.Orientation | ConfigChanges.UiMode | ConfigChanges.ScreenLayout | ConfigChanges.SmallestScreenSize)]
+    public class MainActivity : MauiAppCompatActivity
+    {
+    }
+}
