+++ src/GraphicsControls.Sample/Platforms/Android/MainApplication.cs
