+++ src/GraphicsControls.Sample/Platforms/Android/Resources/values/colors.xml
