+++ src/GraphicsControls.Sample/Platforms/MacCatalyst/AppDelegate.cs
