+++ src/GraphicsControls.Sample/Platforms/MacCatalyst/Info.plist
