+++ src/GraphicsControls.Sample/Platforms/MacCatalyst/Program.cs
@@ -9,7 +9,7 @@ namespace GraphicsControls.Sample
         {
             // if you want to use a different Application Delegate class from "AppDelegate"
             // you can specify it here.
-            UIApplication.Main(args, null, "AppDelegate");
+            UIApplication.Main(args, null, typeof(AppDelegate));
         }
     }
 }
