+++ src/GraphicsControls.Sample/Platforms/Windows/App.xaml
