+++ src/GraphicsControls.Sample/Platforms/Windows/app.manifest
