+++ src/GraphicsControls.Sample/Platforms/iOS/AppDelegate.cs
