+++ src/GraphicsControls.Sample/Platforms/iOS/Info.plist
