+++ src/GraphicsControls.Sample/Platforms/iOS/Resources/LaunchScreen.xib
