+++ src/GraphicsControls.Sample/Resources/Fonts/OpenSans-Regular.ttf
