+++ src/GraphicsControls.Sample/SampleColors.cs
