+++ src/GraphicsControls.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.0.31521.260
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.31829.152
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GraphicsControls", "GraphicsControls\GraphicsControls.csproj", "{474A33AE-06A4-4219-9545-17751EBF3F84}"
 EndProject
@@ -9,9 +9,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Library", "Library", "{4F00
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Sample", "Sample", "{465D03D3-BA73-4E52-BBEA-02DC54647BC6}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GraphicsControls.Sample", "GraphicsControls.Sample\GraphicsControls.Sample\GraphicsControls.Sample.csproj", "{E2B7FB75-A083-4010-835E-C03CF18E18D3}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GraphicsControls.Sample.WinUI", "GraphicsControls.Sample\GraphicsControls.Sample.WinUI\GraphicsControls.Sample.WinUI.csproj", "{02D470CC-4A4F-464D-ACE8-0EC2FCD9E9B4}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GraphicsControls.Sample", "GraphicsControls.Sample\GraphicsControls.Sample.csproj", "{D3CAFC06-FC9C-4703-AD3E-47848DF19696}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -35,45 +33,31 @@ Global
 		{474A33AE-06A4-4219-9545-17751EBF3F84}.Release|iPhone.Build.0 = Release|Any CPU
 		{474A33AE-06A4-4219-9545-17751EBF3F84}.Release|iPhoneSimulator.ActiveCfg = Release|Any CPU
 		{474A33AE-06A4-4219-9545-17751EBF3F84}.Release|iPhoneSimulator.Build.0 = Release|Any CPU
-		{E2B7FB75-A083-4010-835E-C03CF18E18D3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E2B7FB75-A083-4010-835E-C03CF18E18D3}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E2B7FB75-A083-4010-835E-C03CF18E18D3}.Debug|Any CPU.Deploy.0 = Debug|Any CPU
-		{E2B7FB75-A083-4010-835E-C03CF18E18D3}.Debug|iPhone.ActiveCfg = Debug|Any CPU
-		{E2B7FB75-A083-4010-835E-C03CF18E18D3}.Debug|iPhone.Build.0 = Debug|Any CPU
-		{E2B7FB75-A083-4010-835E-C03CF18E18D3}.Debug|iPhone.Deploy.0 = Debug|Any CPU
-		{E2B7FB75-A083-4010-835E-C03CF18E18D3}.Debug|iPhoneSimulator.ActiveCfg = Debug|Any CPU
-		{E2B7FB75-A083-4010-835E-C03CF18E18D3}.Debug|iPhoneSimulator.Build.0 = Debug|Any CPU
-		{E2B7FB75-A083-4010-835E-C03CF18E18D3}.Debug|iPhoneSimulator.Deploy.0 = Debug|Any CPU
-		{E2B7FB75-A083-4010-835E-C03CF18E18D3}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E2B7FB75-A083-4010-835E-C03CF18E18D3}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E2B7FB75-A083-4010-835E-C03CF18E18D3}.Release|Any CPU.Deploy.0 = Release|Any CPU
-		{E2B7FB75-A083-4010-835E-C03CF18E18D3}.Release|iPhone.ActiveCfg = Release|Any CPU
-		{E2B7FB75-A083-4010-835E-C03CF18E18D3}.Release|iPhone.Build.0 = Release|Any CPU
-		{E2B7FB75-A083-4010-835E-C03CF18E18D3}.Release|iPhone.Deploy.0 = Release|Any CPU
-		{E2B7FB75-A083-4010-835E-C03CF18E18D3}.Release|iPhoneSimulator.ActiveCfg = Release|Any CPU
-		{E2B7FB75-A083-4010-835E-C03CF18E18D3}.Release|iPhoneSimulator.Build.0 = Release|Any CPU
-		{E2B7FB75-A083-4010-835E-C03CF18E18D3}.Release|iPhoneSimulator.Deploy.0 = Release|Any CPU
-		{02D470CC-4A4F-464D-ACE8-0EC2FCD9E9B4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{02D470CC-4A4F-464D-ACE8-0EC2FCD9E9B4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{02D470CC-4A4F-464D-ACE8-0EC2FCD9E9B4}.Debug|Any CPU.Deploy.0 = Debug|Any CPU
-		{02D470CC-4A4F-464D-ACE8-0EC2FCD9E9B4}.Debug|iPhone.ActiveCfg = Debug|Any CPU
-		{02D470CC-4A4F-464D-ACE8-0EC2FCD9E9B4}.Debug|iPhone.Build.0 = Debug|Any CPU
-		{02D470CC-4A4F-464D-ACE8-0EC2FCD9E9B4}.Debug|iPhoneSimulator.ActiveCfg = Debug|Any CPU
-		{02D470CC-4A4F-464D-ACE8-0EC2FCD9E9B4}.Debug|iPhoneSimulator.Build.0 = Debug|Any CPU
-		{02D470CC-4A4F-464D-ACE8-0EC2FCD9E9B4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{02D470CC-4A4F-464D-ACE8-0EC2FCD9E9B4}.Release|Any CPU.Build.0 = Release|Any CPU
-		{02D470CC-4A4F-464D-ACE8-0EC2FCD9E9B4}.Release|iPhone.ActiveCfg = Release|Any CPU
-		{02D470CC-4A4F-464D-ACE8-0EC2FCD9E9B4}.Release|iPhone.Build.0 = Release|Any CPU
-		{02D470CC-4A4F-464D-ACE8-0EC2FCD9E9B4}.Release|iPhoneSimulator.ActiveCfg = Release|Any CPU
-		{02D470CC-4A4F-464D-ACE8-0EC2FCD9E9B4}.Release|iPhoneSimulator.Build.0 = Release|Any CPU
+		{D3CAFC06-FC9C-4703-AD3E-47848DF19696}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D3CAFC06-FC9C-4703-AD3E-47848DF19696}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D3CAFC06-FC9C-4703-AD3E-47848DF19696}.Debug|Any CPU.Deploy.0 = Debug|Any CPU
+		{D3CAFC06-FC9C-4703-AD3E-47848DF19696}.Debug|iPhone.ActiveCfg = Debug|Any CPU
+		{D3CAFC06-FC9C-4703-AD3E-47848DF19696}.Debug|iPhone.Build.0 = Debug|Any CPU
+		{D3CAFC06-FC9C-4703-AD3E-47848DF19696}.Debug|iPhone.Deploy.0 = Debug|Any CPU
+		{D3CAFC06-FC9C-4703-AD3E-47848DF19696}.Debug|iPhoneSimulator.ActiveCfg = Debug|Any CPU
+		{D3CAFC06-FC9C-4703-AD3E-47848DF19696}.Debug|iPhoneSimulator.Build.0 = Debug|Any CPU
+		{D3CAFC06-FC9C-4703-AD3E-47848DF19696}.Debug|iPhoneSimulator.Deploy.0 = Debug|Any CPU
+		{D3CAFC06-FC9C-4703-AD3E-47848DF19696}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D3CAFC06-FC9C-4703-AD3E-47848DF19696}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D3CAFC06-FC9C-4703-AD3E-47848DF19696}.Release|Any CPU.Deploy.0 = Release|Any CPU
+		{D3CAFC06-FC9C-4703-AD3E-47848DF19696}.Release|iPhone.ActiveCfg = Release|Any CPU
+		{D3CAFC06-FC9C-4703-AD3E-47848DF19696}.Release|iPhone.Build.0 = Release|Any CPU
+		{D3CAFC06-FC9C-4703-AD3E-47848DF19696}.Release|iPhone.Deploy.0 = Release|Any CPU
+		{D3CAFC06-FC9C-4703-AD3E-47848DF19696}.Release|iPhoneSimulator.ActiveCfg = Release|Any CPU
+		{D3CAFC06-FC9C-4703-AD3E-47848DF19696}.Release|iPhoneSimulator.Build.0 = Release|Any CPU
+		{D3CAFC06-FC9C-4703-AD3E-47848DF19696}.Release|iPhoneSimulator.Deploy.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{474A33AE-06A4-4219-9545-17751EBF3F84} = {4F0017DB-8A18-4153-A129-782B29DECF51}
-		{E2B7FB75-A083-4010-835E-C03CF18E18D3} = {465D03D3-BA73-4E52-BBEA-02DC54647BC6}
-		{02D470CC-4A4F-464D-ACE8-0EC2FCD9E9B4} = {465D03D3-BA73-4E52-BBEA-02DC54647BC6}
+		{D3CAFC06-FC9C-4703-AD3E-47848DF19696} = {465D03D3-BA73-4E52-BBEA-02DC54647BC6}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {475B4FFA-17B6-449C-9D09-DE4E2B0D2F4B}
