+++ src/GraphicsControls/GraphicsControls.csproj
@@ -1,18 +1,19 @@
 <Project Sdk="MSBuild.Sdk.Extras">
 	<PropertyGroup>
-    <TargetFrameworks>net6.0-ios;net6.0-android;net6.0-maccatalyst;net6.0-windows10.0.19041</TargetFrameworks>
-    <RootNamespace>Microsoft.Maui.Graphics.Controls</RootNamespace>	
-    <AssemblyName>Microsoft.Maui.Graphics.Controls</AssemblyName>
-    <LangVersion>8.0</LangVersion>
-    <Nullable>enable</Nullable>
-    <UseMaui>true</UseMaui>
-    <SingleProject>true</SingleProject>
-  </PropertyGroup>
-    <PropertyGroup>
-	   <PackageId>Microsoft.Maui.Graphics.Controls</PackageId>
+		<TargetFrameworks>net6.0-ios;net6.0-android;net6.0-maccatalyst</TargetFrameworks>
+		<TargetFrameworks Condition="$([MSBuild]::IsOSPlatform('windows')) and '$(MSBuildRuntimeType)' == 'Full'">$(TargetFrameworks);net6.0-windows10.0.19041</TargetFrameworks>
+		<RootNamespace>Microsoft.Maui.Graphics.Controls</RootNamespace>
+		<AssemblyName>Microsoft.Maui.Graphics.Controls</AssemblyName>
+		<LangVersion>8.0</LangVersion>
+		<Nullable>enable</Nullable>
+		<UseMaui>true</UseMaui>
+		<SingleProject>true</SingleProject>
 	</PropertyGroup>
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Maui.Graphics" Version="6.0.101-preview.10.640" />
-  </ItemGroup>
-  <Import Project="..\..\Microsoft.Maui.Graphics.Controls.MultiTargeting.targets" />
+	<PropertyGroup>
+		<PackageId>Microsoft.Maui.Graphics.Controls</PackageId>
+	</PropertyGroup>
+	<ItemGroup>
+		<PackageReference Include="Microsoft.Maui.Graphics" Version="6.0.101-preview.10.640" />
+	</ItemGroup>
+	<Import Project="..\..\Microsoft.Maui.Graphics.Controls.MultiTargeting.targets" />
 </Project>
